Flora with manual YSorting + shitload of things
Added background particles (child of player)

Fixed level scripts

Created new scene for flora. This is a more versatile version of the Tree scene. Its export variables are interactive in the editor. Its script uses the player position to determine its Z layer. To save resources, there is an exported variable called Reachable which dictates wheather or not the object can be reached by the player. If it can't, there is no need to change its Z value. When Tree is selected, there are particles. <3

Because the player can now walk between the trees, there were changes to the bounds of the first island in the red level.

Because of the changes to the bounds of the first island, it was possible to skip the first crystal. Changes were made to the platforms between the red and the orange levels to prevent that without making the game too hard.

Crystals are also ordered in the Z layer. The press_e node is now inside the crystal scene and its Z is also affected by player position.

Added a node called spawn_location to the checkpoint scene, and now this is where the player will spawn. This is good to keep the flow of the game.

CHANGED EVERY FREAKING CHECKPOINT IN THE GAME BECAUSE OF THE LAST CHANGE.

The deer has a separate scene now.